[#138] feat: 지난달 소비내역, 페어링 답변 데이터 가공함수 분리
